requirements.txt
streamlit
docxtpl
python-docx
qrcode
pandas
aspose-words-cloud
